src/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.0.0
python-dotenv==1.0.1
pydantic==2.9.0
python-multipart==0.0.9
bcrypt==4.1.2


# JWT for auth
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.3

